==> 30_TX_data_preprocessing.ipynb <==
# Import libraries and data

import pandas as pd
import os
import numpy as np
import plotly.graph_objs as go
import plotly.express as px
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
warnings.filterwarnings("ignore")

import pickle
with open("TX_data/variables.pickle", "rb") as f:
    me_2017,me_2018,me_2019,me_2021,me_2022,df_metric_2017,df_metric_2018,df_metric_2019,df_metric_2021,df_metric_2022,df_me_2017,df_me_2018,df_me_2019,df_me_2021,df_me_2022,metric_year,df3= pickle.load(f)

with open("TX_data/variables2.pickle", "rb") as f:
    df_me = pickle.load(f)

# Data Preprocess

## 1. abbreviations_explanation

# """
# Organize the meanings of abbreviations
# """
# abb=pd.read_csv('TX_data/abbreviations_explanation.csv')
# df_split = abb['meaning'].str.split('(', expand=True)
# abb['meaning']=df_split[0]
# abb['abbr'] = abb['abbr'].combine_first(df_split[1].str.replace(')', '') )
# abb.to_csv('/Users/dai/Desktop/abbreviations.csv')

abb=pd.read_csv('TX_data/abbreviations.csv')

## 2. Aggregating TAPR_District Reports by Year

folder_path = 'TX_data/TAPR_District/'

for year in [2017, 2018,2019,2021,2022]: 
    i=0
    merged_df = pd.DataFrame()
    
    subfolder_path = os.path.join(folder_path, str(year))
    
    for filename in os.listdir(subfolder_path):
         
        if filename.endswith('.csv')  or filename.endswith('.xls'):
            file_path = os.path.join(subfolder_path, filename)

            if filename.endswith('.csv'):
                df = pd.read_csv(file_path, index_col=0)
            elif filename.endswith('.xls'):
                df = pd.read_html(file_path)[0]
            if i==0:
                merged_df = pd.concat([merged_df, df])
            else:
                merged_df = pd.merge(merged_df, df,how='left',on=['DISTRICT'])
        i+=1
            
    merged_df.to_csv(folder_path+'TAPR_'+str(year)+'.csv')

## 3. Pre-process house price data

Note:

1. Please get `zip_code_market_tracker.tsv000`  file first:

    [Average home price by zip code](https://www.redfin.com/news/data-center/): Scroll down to “How it Works” section, click hyperlink “Zip Code”

2. put it into `TX_data` folder.



dh = pd.read_csv('TX_data/zip_code_market_tracker.tsv000', sep='\t')

# Select the desired area
dh=dh[dh['state']=='Texas']
dh['period_begin'] = pd.to_datetime(dh['period_begin'])
dh['period_end'] = pd.to_datetime(dh['period_end'])

dh['area']=dh['median_sale_price']/dh['median_ppsf']
dh['region']=dh['region'].str.replace('Zip Code: ','')
dh.rename(columns={'region':'zipcode'},inplace=True)
dh['zipcode'] = dh['zipcode'].astype(str)

dh=dh.sort_values('period_begin').reset_index(drop=True)
dh=dh.sort_values(['zipcode','period_begin','period_end','property_type_id'])

dh.head(3)

# Choose related field
dh1=dh.loc[dh['property_type']=='All Residential',['zipcode', 'period_end','median_ppsf','area']]

# take last year's dec data as the next year data
## eg. 2022-12-01~2023-02-28 is considered as 2023
dh1=dh1[dh1['period_end'].isin(['2015-12-31','2016-12-31','2017-12-31','2018-12-31','2019-12-31','2020-12-31','2021-12-31','2022-12-31'])]
dh1['year'] = dh1['period_end'].dt.year

# calculate change percent
group = dh1.groupby('zipcode')
dh1['median_ppsf_prev'] = group['median_ppsf'].shift(1)
dh1['median_ppsf_change_percent'] = (dh1['median_ppsf'] - dh1['median_ppsf_prev']) / dh1['median_ppsf_prev'] *100
dh1=dh1.dropna(axis=0,how='any') #drop all rows that have any NaN values

dh1.head(3)

# Choose related field -> for all year range
dh2=dh.loc[dh['property_type']=='All Residential',['zipcode', 'period_end','median_ppsf','area']]
dh2['year'] = dh2['period_end'].dt.year
#dh2['year']=dh2['year']+1
dh2=dh2.dropna(axis=0,how='any') #drop all rows that have any NaN values

## 4. Integration of District ID and Zipcode with House Data

Source: 
- https://schoolsdata2-tea-texas.opendata.arcgis.com/ > Archived Schools Data
- https://schoolsdata2-db440-tea-texas.opendata.arcgis.com/


# zipcode dataset
z=pd.read_csv('TX_data/School2020to2021.csv').sort_values('District_1')[[ 'District_1', 'District_N', 'Zip5']].drop_duplicates()
z.rename(columns={'Zip5':'zipcode','District_1':'DISTRICT','District_N':'DISTNAME'},inplace=True)
z['zipcode'] = z['zipcode'].astype(str)

display(z.head(3))

# merge with house price data
dt=pd.merge(dh1, z, on='zipcode', how='left')  #shape change: Because one zipcode may correspond to more distname
dt=dt[dt['DISTRICT'].notna()]

display(dt.head(3))

# save data
dt.to_csv('TX_data/HousePrice/TX_houseprice_2016to2023.csv',index=False)

dt2=pd.merge(dh2, z, on='zipcode', how='left')  #shape change: Because one zipcode may correspond to more distname
dt2=dt2[dt2['DISTRICT'].notna()]

# save data
dt2.to_csv('TX_data/HousePrice/TX_houseprice.csv',index=False)

## 5. Merge above data by year

# creat new df by year
for year, group in dt.groupby('year'):
    globals()[f"dt_{year}"] = group.copy()
    globals()[f"dt_{year}"].to_csv('TX_data/HousePrice/house'+str(year)+'.csv')

# merge
for year in [2017, 2018,2019,2021,2022]: 
    globals()[f"TAPR_{year}"] = pd.read_csv('TX_data/TAPR_District/TAPR_'+str(year)+'.csv',na_values=['.'])
    globals()[f"TAPR_{year}"]['DISTRICT']=globals()[f"TAPR_{year}"]['DISTRICT'].str.replace("'",'').astype('float64')
    if 'Unnamed: 0' in globals()[f"TAPR_{year}"].columns:
        globals()[f"TAPR_{year}"]= globals()[f"TAPR_{year}"].drop(columns=['Unnamed: 0'])
    globals()[f"merged_{year}"] = pd.merge(globals()[f"dt_{year-1}"], globals()[f"TAPR_{year}"], on='DISTRICT', how='inner')

    # rename first DISTNAME_x
    col_names = list(globals()[f"merged_{year}"].columns)
    idx = col_names.index('DISTNAME_x')
    col_names[idx] = 'distname'
    globals()[f"merged_{year}"].columns = col_names
    
    # drop duplicate column 'DISTNAME_'
    drop_list=globals()[f"merged_{year}"].columns[globals()[f"merged_{year}"] .columns.str.startswith('DISTNAME')]
    globals()[f"merged_{year}"] = globals()[f"merged_{year}"].drop(columns=drop_list)
    
    # save data
    globals()[f"merged_{year}"].to_csv('TX_data/merged_'+str(year)+'.csv')

# Functions

## 1. Map District Names to their Corresponding IDs

# Input: a list of district name (default: all exist district name)
# Output: a dictionary, district_name: districtid.
def find_districtid(district_list=None):
    if district_list:
        did=pd.read_csv('TX_data/Districts2020to2021.csv')[['NAME','DISTRICT']]
        districtid = {}
        for d in district_list:
            id=did.loc[did['NAME']==d,'DISTRICT'].values[0]
            districtid[d] = id.replace('-','')
        #{'Dallas ISD': '057905', 'Fort Worth ISD': '220905'}
    else:
        did=pd.read_csv('TX_data/Districts2020to2021.csv')[['NAME','DISTRICT']]
        did['DISTRICT']=did['DISTRICT'].str.replace('-','')
        districtid=dict(zip(did['NAME'], did['DISTRICT']))
    return districtid

## 2. Retrieving Relevant IDs for a Given Field


def find_element(Data_field, Data_Level,Subcategory=None, year=None,detail='Y'):
    file_path = 'TX_data/datadict/district_processed/'
    if year is not None:
        csv_files = [os.path.join(file_path,'datadict_district_'+str(year)+'.csv' ),
                     os.path.join(file_path,'datadict_district_'+str(year-1)+'.csv' )]
    else:
        csv_files = [os.path.join(file_path, f) for f in os.listdir(file_path) if f.endswith(".csv") and f.startswith('datadict')]

    elements=[]
    for csv_file in csv_files:
        output = pd.read_csv(csv_file)
        
        # output = output.loc[output.iloc[:, 0] == Data_field, :]
        output = output.loc[(output.iloc[:, 0].str.startswith(Data_field)) & (~output.iloc[:, 0].isna()), :]
        
        if Subcategory is not None:
            #output=output.loc[ output.iloc[:,1]==Subcategory ,:]
            output=output.loc[ output.iloc[:,1].str.startswith(Subcategory) ,:]

        if year is not None:
            output=output.loc[ output.iloc[:,2]==str(year),: ]

        element = output.loc[:, Data_Level]
        if not element.empty:
            elements.append(element.values[0])
    
    print('The relevant element ids are: ',elements,'\n')
    
    if detail=='Y':
        if type(elements)=='str':
            elements=[elements]
        for enum in elements:
            try:
                position = df_me[df_me['elements'].apply(lambda x: enum in x)].index.tolist()[0]
            except:
                print('please make sure ', enum,' not belong to State and Region ## column')
                
            e_years=df_me.loc[position,'occurrence_year']
            e_metric=df_me.loc[position,'metric']
            meaning=df_me.loc[position,'meaning']
            print(enum,' belongs to metric: ',e_metric) 
            print('Occurrence year: ', e_years)
            print("Meaning: ",meaning)  
            print('\n')      
                
    return elements

## 3. District Dataframe Retrieval Function: Explaining its Elements and Usage

def select_columns(df, columns):
    existing_columns = set(df.columns)
    selected_columns = [col for col in columns if col in existing_columns]
    return df[selected_columns]

def mapping5_list(enum_list):
    path_list=[]
    ae=[] 
    if type(enum_list)=='str':
        enum_list=[enum_list]
    for enum in enum_list:
        try:
            position = df_me[df_me['elements'].apply(lambda x: enum in x)].index.tolist()[0]
        except:
            print('please make sure ', enum,' not belong to State and Region ## column')
            continue
            
        e_years=df_me.loc[position,'occurrence_year']
        e_metric=df_me.loc[position,'metric']
        
        # column name
        for ind,i in enumerate(e_years):
            ae.append(enum[:-3] + str(i)[2:4] + enum[-1])
            ae.append(enum[:-3] + str(i-1)[2:4] + enum[-1])
            ae.append(enum[:-3] + str(i-2)[2:4] + enum[-1])
            path_list.append(os.path.join('TX_data','TAPR_District',str(i), e_metric))
            
    ae=list(set(ae))
    ae.sort()
    ae.insert(0,'DISTRICT')
    path_list=list(set(path_list))
    path_list.sort()
        
    for ind,path in enumerate(path_list):
        if int(path.split('/')[2])<=2019:
            df = pd.read_html(path+'.xls')[0]
        else:
            df = pd.read_csv(path+'.csv')
        dc= select_columns(df, ae)
        
        if ind==0:
            output=dc
        else:
            output=pd.merge(output,dc,on='DISTRICT',how='outer')
    
    # Deal with duplicate columns. Save the one occur in later year.
    output = output[output.columns[~output.columns.str.endswith('_x')]]
    output=output.rename(columns={col: col.rstrip('_y') for col in output.columns if col.endswith('_y')})
    output.sort_index(axis=1, ascending=False, inplace=True)
    
    # Sort columns by its name
    cols = output.columns.tolist()
    cols.sort()
    cols.remove('DISTRICT')
    cols.insert(0, 'DISTRICT')
    output = output[cols]
    output.iloc[:,0]=output.iloc[:,0].str.strip("'")
    output = output.set_index(output.columns[0])
    return output

## Running Example

# 1. give spefic district name, return its id
districtid0=find_districtid(district_list=None)
print(districtid0)

district_list=['Dallas ISD','Fort Worth ISD']
districtid=find_districtid(district_list)
print(districtid)

# 2.1 give wanted field
test=pd.read_csv('TX_data/5years_trends_try.csv')
Data_field=test['Data field'][0]
Subcategory	= test['Subcategory'][0]
Data_Level=test['Data Level'][0]
print('The field we want is: '+Data_field,Subcategory,Data_Level,sep='; ')

# 2.2 return element id
enum_list=find_element(Data_field,Data_Level, Subcategory, year=None,detail='Y')  

# 3. return the result dataframe. 
if enum_list != None:
    dt=mapping5_list(enum_list)

dt=dt.loc[list(districtid.values()), :]  #bc: the number of `districtid`` less than actual `district``
dt = dt.replace('.', np.nan).astype(float)
new_rows = {v: k for k, v in districtid.items()}  
dt = dt.rename(index=new_rows)
display(dt)

# General EDA

def factor_name(element):
    year=element[-3:-1]
    file_path = 'TX_data/datadict/district_processed/'
    csv_files = [os.path.join(file_path, f) for f in os.listdir(file_path) if f.endswith('.csv') and str(year) in f][0]
    df2=pd.read_csv(csv_files)
    mask = df2.where(df2 == element)
    stacked = mask.stack()
    info=', '.join([df2.iloc[stacked.index.get_level_values(0), 0].values[0],df2.iloc[stacked.index.get_level_values(0), 1].values[0], stacked.index.get_level_values(1).values[0], str(year)])
    info_dict={element:info}
    return info_dict
## eg. element = 'DDA00AR01218R'

def analysis(df,top=None):
    
    df_mean = df.mean(axis=1) 
    if top!=None:
        df_sorted = df.mean().sort_values(ascending=False)
        cols_top = df_sorted.index[:top]
        df = df[cols_top]
    
    df_mean_select = df.mean(axis=1)
    
    fig = go.Figure()
    
    # Add scatter plot for each column
    fig = px.scatter(df, x=df.index, y=df.columns, title='Scatter Plot')
    fig.add_trace(go.Scatter(x=df.index, y=df_mean_select, name='Select average', mode='markers', line=dict(color='black', width=4)))
    fig.add_trace(go.Scatter(x=df.index, y=df_mean, name='Overall average', mode='markers', line=dict(color='grey', width=4)))
    fig.show()
    
    for col in df.columns:
        fig.add_trace(go.Scatter(x=df.index, y=df[col], name=col))
    fig.add_trace(go.Scatter(x=df.index, y=df_mean_select, name='Select average', line=dict(color='black', width=3)))
    fig.add_trace(go.Scatter(x=df.index, y=df_mean, name='Overall average', line=dict(color='grey', width=3)))
    fig.update_layout(title='Line Plot', xaxis_title='X Axis', yaxis_title='Y Axis')
    fig.show()

    # Add box plot for each column
    fig = px.box(df, y=df.columns, title='Box Plot')
    fig.add_trace(go.Box(y=df_mean_select, name='Select average', line=dict(color='black', width=3)))
    fig.add_trace(go.Box(y=df_mean, name='Overall average', line=dict(color='grey', width=3)))
    fig.show()

    # Add pct change plot for each column
    df_pct_change = df.pct_change()
    df_pct_change_long = df_pct_change.melt(value_name='pct_change', var_name='column')
    fig = px.line(df_pct_change_long, x='column', y='pct_change', title='Pct Change Trend Analysis')
    fig.show()

df=dt.T.copy()

info_dict = {}
for element in df.index:
    info_dict.update(factor_name(element))

for key, value in info_dict.items():
    print(f"{key}: {value}")
    
analysis(df,top=30)
